Remove trailing semicolons to force identical return types
While not generally a problem with VkResult because of `#[must_use]`
removing the trailing semicolon at the end of a function body enforces
that the last expression - in most cases a virtual call to the loaded
function-ptr - retuns exactly the same type. In most cases that is void
(`()`), and if that changes we'll get a compile error instead of it
going unnoticed.